+++ src/app-config/dicts.ts
@@ -250,6 +250,47 @@ export function getALlDicts () {
         cnfirst: false
       }
     },
+    googledict: {
+      /**
+       * Supported language: en, zh-CN, zh-TW
+       * `1` for supported
+       */
+      lang: '111',
+      /**
+       * Full content page to jump to when user clicks the title.
+       * %s will be replaced with the current word.
+       * %z will be replaced with the traditional Chinese version of the current word.
+       * %h will be replaced with the current word joining with hyphen "-".
+       */
+      page: 'https://www.google.com.hk/search?q=define+%s',
+      /**
+       * If set to true, the dict start searching automatically.
+       * Otherwise it'll only start seaching when user clicks the unfold button.
+       * Default MUST be true and let user decide.
+       */
+      defaultUnfold: true,
+      /**
+       * This is the default height when the dict first renders the result.
+       * If the content height is greater than the preferred height,
+       * the preferred height is used and a mask with a view-more button is shown.
+       * Otherwise the content height is used.
+       */
+      preferredHeight: 240,
+      /** Word count to start searching */
+      selectionWC: {
+        min: 1,
+        max: 5,
+      },
+      /** Only start searching if the selection contains the language. */
+      selectionLang: {
+        eng: true,
+        chs: true,
+      },
+      /** Optional dict custom options. Can only be boolean or number. */
+      options: {
+        enresult: true
+      }
+    },
     guoyu: {
       /**
        * Supported language: en, zh-CN, zh-TW