+++ src/components/dictionaries/longman/View.tsx
@@ -2,8 +2,50 @@ import React from 'react'
 import Speaker from '@/components/Speaker'
 import StarRates from '@/components/StarRates'
 import { LongmanResult, LongmanResultLex, LongmanResultRelated, LongmanResultEntry } from './engine'
+import { message } from '@/_helpers/browser-api'
+import { MsgType, MsgAudioPlay } from '@/typings/message'
 
 export default class DictLongman extends React.PureComponent<{ result: LongmanResult }> {
+  _audioDelayTimeout: any
+
+  isAudioElement (evt: React.MouseEvent<HTMLDivElement>): boolean {
+    const target = (evt.target as HTMLElement)
+    const cls = target.classList
+    return cls && cls.contains('dictLongman-Speaker')
+  }
+
+  handleDictMouseOver = (evt: React.MouseEvent<HTMLDivElement>) => {
+    if (this.isAudioElement(evt)) {
+      clearTimeout(this._audioDelayTimeout)
+      // React resuses synthetic event object
+      const target = evt.target as HTMLElement
+      this._audioDelayTimeout =
+        setTimeout(() => this.playAudio(target), 400)
+    }
+  }
+
+  handleDictMouseOut = (evt: React.MouseEvent<HTMLDivElement>) => {
+    if (this.isAudioElement(evt)) {
+      clearTimeout(this._audioDelayTimeout)
+    }
+  }
+
+  handleDictClick = (evt: React.MouseEvent<HTMLDivElement>) => {
+    if (this.isAudioElement(evt)) {
+      clearTimeout(this._audioDelayTimeout)
+      const target = evt.target as HTMLElement
+      target.blur()
+      this.playAudio(target)
+    }
+  }
+
+  playAudio = (target: HTMLElement) => {
+    const src = target.dataset.srcMp3
+    if (src) {
+      message.send<MsgAudioPlay>({ type: MsgType.PlayAudio, src })
+    }
+  }
+
   renderEntry (entry: LongmanResultEntry) {
     return (
       <section key={entry.title.HWD + entry.title.HOMNUM} className='dictLongman-Entry'>
@@ -14,7 +56,15 @@ export default class DictLongman extends React.PureComponent<{ result: LongmanRe
               <span className='dictLongman-Title_HYPHENATION'>{entry.title.HYPHENATION}</span>
               <span className='dictLongman-Title_HOMNUM'>{entry.title.HOMNUM}</span>
             </h1>
-            {entry.level ? <StarRates rate={entry.level} className='dictLongman-Level' /> : null}
+            {
+              entry.level ?
+              (
+                <span title={entry.level.title} className='dictLongman-Level'>
+                  <StarRates max={3} rate={entry.level.rate} className='dictLongman-Level_Rate' />
+                </span>
+              )
+              : null
+            }
             {entry.freq && entry.freq.map(freq => (
               <span key={freq.rank} className='dictLongman-FREQ' title={freq.title}>{freq.rank}</span>
             ))}
@@ -79,9 +129,14 @@ export default class DictLongman extends React.PureComponent<{ result: LongmanRe
           <div className='dictLongman-Wordfams' dangerouslySetInnerHTML={{ __html: result.wordfams }} />
         }
 
-        {dicts.map(dict => result[dict].length > 0
+        {dicts.map((dict, index) => result[dict].length > 0
           ? (
-            <div className='dictLongman-Dict'>
+            <div className='dictLongman-Dict'
+              key={dict + index}
+              onClick={this.handleDictClick}
+              onMouseOver={this.handleDictMouseOver}
+              onMouseOut={this.handleDictMouseOut}
+            >
               {/* <h1 className='dictLongman-DictTitle'>
                 <span>- {dictTitle[dict]} -</span>
               </h1> */}