+++ src/components/dictionaries/longman/engine.ts
@@ -21,7 +21,10 @@ export interface LongmanResultEntry {
     href: string
   }
   phsym?: string
-  level: number
+  level?: {
+    rate: number,
+    title: string
+  }
   freq?: Array<{
     title: string
     rank: string
@@ -86,6 +89,16 @@ function handleDOMLex (
 
   const audio: { uk?: string, us?: string } = {}
 
+  doc.querySelectorAll<HTMLSpanElement>('.speaker.exafile').forEach(
+    $speaker => {
+      const mp3 = $speaker.dataset.srcMp3
+      if (mp3) {
+        $speaker.outerHTML =
+          `<button data-src-mp3="${mp3}" title="${$speaker.title}" class="dictLongman-Speaker">🔊</button>`
+      }
+    }
+  )
+
   if (options.wordfams) {
     result.wordfams = getInnerHTML(doc, '.wordfams')
   }
@@ -116,7 +129,6 @@ function handleDOMLex (
       },
       prons: [],
       senses: [],
-      level: 0,
     }
 
     const $topic = $entry.querySelector<HTMLAnchorElement>('a.topic')
@@ -136,8 +148,18 @@ function handleDOMLex (
 
     entry.phsym = getText($head, '.PronCodes')
 
-    const level = getText($head, '.LEVEL')
-    entry.level = (level.match(/●/g) || []).length
+    const $level = $head.querySelector('.LEVEL') as HTMLSpanElement
+    if ($level) {
+      const level = {
+        rate: 0,
+        title: ''
+      }
+
+      level.rate = (($level.textContent || '').match(/●/g) || []).length
+      level.title = $level.title
+
+      entry.level = level
+    }
 
     entry.freq = Array.from($head.querySelectorAll<HTMLSpanElement>('.FREQ'))
       .map($el => ({