+++ src/typings/message.ts
@@ -53,11 +53,13 @@ export const enum MsgType {
   /** Query panel state */
   QueryPanelState,
 
+  /** Manually emit selection event */
+  EmitSelection,
+
   /**
    * Background proxy sends back underlyingly
    */
   __PageInfo__,
-  __PreloadSelection__,
 }
 
 /** iframe messaging */
@@ -72,6 +74,7 @@ export interface MsgSelection {
   readonly mouseY: number
   readonly dbClick?: boolean
   readonly ctrlKey?: boolean
+  /** skip salad bowl and show panel directly */
   readonly instant?: boolean
   /** force panel to skip reconciling position */
   readonly force?: boolean